DSP-23519. Build on ubuntu:18.04
Rationale for this change is to have more recent versions of libraries
included in the final tarbal build using the Dockerfile.
Especially openssl related: libcrypto.so and libssl.so.

Especially for openssl:
16.04 is using 1.0.2g while 18.04 uses 1.1.1-11.

Broader context is that DSE is including extracted tarball
which then causes security scanners to find libs from collectd
such as libcrypto.so to be source of CVEs.